intel_th: Support NPKT ACPI table in MSU
When the bios is configured to trace to memory, it will
 * trace to MTB before memory is available
 * save MTB traces to a memory area when memory is available
 * trace to CSR in a reserved memory area
In case of reboot, MTB and CSR buffers may be backed up by bios in
recovery copies.

The memory area for all those buffers is exposed to the kernel through
NPKT ACPI table.

Stop the current trace if the CSR capture is still in progress when the
user either tries to read the CSR buffer or starts a new capture.

Make the various buffers available to userspace through debugfs.

Signed-off-by: Yann Fouassier <yann.fouassier@intel.com>
Signed-off-by: Laurent FERT <laurent.fert@intel.com>
Signed-off-by: Tian, Baofeng <baofeng.tian@intel.com>