Revert "mfd: intel-lpss: Avoid resuming runtime-suspended lpss unnecessarily"
This reverts commit d4688de00e18f1286122957e5f868c3091156bc3.

It is causing warnings from clk_core_disable() and clk_core_unprepare()
in drivers/i2c/busses/i2c-designware-platdrv.c due to a possibility to
enter suspend callback twice.

We don't have a clear or accepted fix for that. Best candidate so far is
unfortunately causing regression to previous platforms so it's better to
hold the patch before those are reworked.

Fixes: LCK-3500
Signed-off-by: Jarkko Nikula <jarkko.nikula@linux.intel.com>
Signed-off-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>